all: unify test files
