--- base_json_test.go
@@ -1,25 +1,0 @@
-// SPDX-FileCopyrightText: 2021 The Go Language Server Authors
-// SPDX-License-Identifier: BSD-3-Clause
-
-//go:build !gojay
-// +build !gojay
-
-package protocol
-
-import (
-	"testing"
-
-	"github.com/segmentio/encoding/json"
-)
-
-func TestCancelParams(t *testing.T) {
-	t.Parallel()
-
-	testCancelParams(t, json.Marshal, json.Unmarshal)
-}
-
-func TestProgressParams(t *testing.T) {
-	t.Parallel()
-
-	testProgressParams(t, json.Marshal, json.Unmarshal)
-}
