--- selectionrange_json.go
@@ -1,10 +1,0 @@
-// SPDX-FileCopyrightText: 2021 The Go Language Server Authors
-// SPDX-License-Identifier: BSD-3-Clause
-
-//go:build !gojay
-// +build !gojay
-
-package protocol
-
-// SelectionRangeProviderOptions selection range provider options interface.
-type SelectionRangeProviderOptions interface{}
